requirements.txt
amqp==2.1.4
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.7
astroid==2.3.3
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.2
celery==4.0.2
certifi==2020.4.5
chardet==3.0.4
colorama==0.4.3
decorator==4.0.11
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.5
django-appconf==1.0.4
django-filter==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
isort==4.3.21
kombu==4.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
pilkit==2.0
Pillow==7.1.1
pipreqs==0.4.10
psycopg2==2.8.5
pylint==2.4.4
pytz==2017.2
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
vine==1.1.3
virtualenv==20.0.16
virtualenv-clone==0.5.4
whitenoise==5.0.1
wrapt==1.11.2